simplified rotation logic as now I understand how a rotation matrix works in 2D.
the vector c is the sum of vectors a & b.

the rotation of the sum of 2 vectors (a, b) == the rotation of a + the rotation of b == the rotation of c.

Therefor the rotation of vector n ==

the rotation if n's components (x, y) added together

or rot(x) + rot(y) = rot(n)

Note, it is easy to rotation a vector with one axis that is 0, because we can translate it to the circle origin in a unit circle (0 degrees, radius 1)
and calculate the rotation (Y/opposite = sine(angle), X/adjacent = cosine(angle))
remember, SOH CAH TOA: Sine = opp/hyp. Cosine = adj/hyp. Hypotenuse/longest side of this triangle in a unit circle is 1 because the radius is 1.